0030855: Coding - New warnings within Standard_ReadLineBuffer after integration fix for 0030830
Added missing cast.

+++ src/Standard/Standard_ReadLineBuffer.hxx
@@ -172,7 +172,7 @@ protected:
                    size_t theLen,
                    size_t& theReadLen)
   {
-    theReadLen = theStream.read (myReadBuffer.data(), theLen).gcount();
+    theReadLen = (size_t )theStream.read (myReadBuffer.data(), theLen).gcount();
     return !theStream.bad();
   }
